#660 Added the 'keepIdentity' argument and its Integration Tests. (Ref: #372)

+++ RepoDb.Extensions/RepoDb.PostgreSql.BulkOperations/RepoDb.PostgreSql.BulkOperations.IntegrationTests/Helper.cs
@@ -323,7 +323,7 @@ namespace RepoDb.PostgreSql.BulkOperations.IntegrationTests
                 var index = i + 1;
                 tables.Add(new
                 {
-                    Id = hasId ? index : 0,
+                    Id = (long)(hasId ? index : 0),
                     ColumnBigInt = (long)random.Next(100),
                     ColumnBoolean = true,
                     ColumnInteger = random.Next(100),
